+++ tests/test_chat.py
@@ -4,7 +4,7 @@ import unittest
 
 import pytest
 
-from pylangacq.chat import _clean_word, Reader
+from pylangacq.chat import _clean_word, Reader, _CLITIC_REGEX
 from pylangacq.tests.test_chat import (
     BaseTestCHATReader,
     _LOCAL_EVE_PATH,
@@ -41,3 +41,24 @@ def test_if_childes_has_updated_data():
 )
 def test__clean_word(original, expected):
     assert _clean_word(original) == expected
+
+
+@pytest.mark.parametrize(
+    "source, expected_pre, expected_core, expected_post",
+    [
+        ("abc", None, "abc", None),
+        ("abc$def~ghi", "abc", "def", "ghi"),
+        ("blah$abc$def~ghi", "blah$abc", "def", "ghi"),
+        ("abc$def", "abc", "def", None),
+        ("abc$def$blah", "abc$def", "blah", None),
+        ("abc$def~ghi~blah", "abc", "def", "ghi~blah"),
+        ("def~ghi~blah", None, "def", "ghi~blah"),
+        ("blah$abc$def~ghi~blah", "blah$abc", "def", "ghi~blah"),
+    ],
+)
+def test__clitic_regex(source, expected_pre, expected_core, expected_post):
+    match = _CLITIC_REGEX.search(source)
+    _, actual_pre, actual_core, _, actual_post = match.groups()
+    assert actual_pre == expected_pre
+    assert actual_core == expected_core
+    assert actual_post == expected_post